@@ -2,6 +2,7 @@
 
 Run via MCP to print per-part physics properties at frames 1 and 20.
 """
+
 import bpy
 from mathutils import Vector
 
@@ -17,8 +18,8 @@ def inspect_frame(frame: int):
         pass
     deps = bpy.context.evaluated_depsgraph_get()
 
-    collider = bpy.data.objects.get('Sorting_Bucket_Collider')
-    bucket = bpy.data.objects.get('Sorting_Bucket')
+    collider = bpy.data.objects.get("Sorting_Bucket_Collider")
+    bucket = bpy.data.objects.get("Sorting_Bucket")
     top_z = None
     if collider:
         bbox = [collider.matrix_world @ Vector(c) for c in collider.bound_box]
@@ -27,19 +28,28 @@ def inspect_frame(frame: int):
         bbox = [bucket.matrix_world @ Vector(c) for c in bucket.bound_box]
         top_z = max(p.z for p in bbox)
 
-    print(f'--- Frame {frame} ---')
-    print(f'Bucket top Z: {top_z}')
+    print(f"--- Frame {frame} ---")
+    print(f"Bucket top Z: {top_z}")
 
-    parts_coll = bpy.data.collections.get('lego_parts')
+    parts_coll = bpy.data.collections.get("lego_parts")
     if not parts_coll:
-        print('No lego_parts collection')
+        print("No lego_parts collection")
         return
 
     for obj in parts_coll.objects:
-        rb = getattr(obj, 'rigid_body', None)
+        rb = obj.rigid_body
         eval_obj = obj.evaluated_get(deps)
         loc = eval_obj.location
-        print(f'{obj.name}: loc=({loc.x:.3f},{loc.y:.3f},{loc.z:.3f}), rb={bool(rb)}, mass={getattr(rb, "mass", None)}, shape={getattr(rb, "collision_shape", None)}, margin={getattr(rb, "collision_margin", None)}')
+        mass = None
+        shape = None
+        margin = None
+        if rb is not None:
+            mass = getattr(rb, "mass", None)
+            shape = getattr(rb, "collision_shape", None)
+            margin = getattr(rb, "collision_margin", None)
+        print(
+            f"{obj.name}: loc=({loc.x:.3f},{loc.y:.3f},{loc.z:.3f}), rb={bool(rb)}, mass={mass}, shape={shape}, margin={margin}"
+        )
 
 
 def main():
@@ -47,5 +57,5 @@ def main():
     inspect_frame(20)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()
